TST: collect Index tests by method (#40070)
