+++ pandas/tests/indexes/timedeltas/test_timedelta.py
@@ -64,31 +64,6 @@ class TestTimedeltaIndex(DatetimeLike):
             index.isin([index[2], 5]), np.array([False, False, True, False])
         )
 
-    def test_sort_values(self):
-
-        idx = TimedeltaIndex(["4d", "1d", "2d"])
-
-        ordered = idx.sort_values()
-        assert ordered.is_monotonic
-
-        ordered = idx.sort_values(ascending=False)
-        assert ordered[::-1].is_monotonic
-
-        ordered, dexer = idx.sort_values(return_indexer=True)
-        assert ordered.is_monotonic
-
-        tm.assert_numpy_array_equal(dexer, np.array([1, 2, 0]), check_dtype=False)
-
-        ordered, dexer = idx.sort_values(return_indexer=True, ascending=False)
-        assert ordered[::-1].is_monotonic
-
-        tm.assert_numpy_array_equal(dexer, np.array([0, 2, 1]), check_dtype=False)
-
-    def test_argmin_argmax(self):
-        idx = TimedeltaIndex(["1 day 00:00:05", "1 day 00:00:01", "1 day 00:00:02"])
-        assert idx.argmin() == 1
-        assert idx.argmax() == 0
-
     def test_misc_coverage(self):
 
         rng = timedelta_range("1 day", periods=5)
